Refactored Code. Removed redundant decorator, added equals and hashcode for searchNode
