+++ statespace/src/StateSpaceSearch/Goal.java
@@ -1,0 +1,5 @@
+package StateSpaceSearch;
+
+public abstract class Goal<State> {
+    public abstract Boolean satisfied(State state);
+}
