+++ statespace/src/StateSpaceSearch/Search.java
@@ -1,22 +1,22 @@
-import java.util.*;
+package StateSpaceSearch;
+
+import java.util.List;
 
 public class Search<State> {
-    // constructor
-    public Search(State startState, Goal<State> goal, SearchController<State> controller,
-            Successors<State> successors) {
+
+    public Search(State startState, Goal<State> goal, Controller<State> controller,
+                  Successors<State> successors) {
         this.goal = goal;
         this.controller = controller;
         this.successors = successors;
         this.startState = startState;
     }
 
-    // private members
     private Goal<State> goal;
-    private SearchController<State> controller;
+    private Controller<State> controller;
     private Successors<State> successors;
     private State startState;
 
-    // methods
     public SearchPath<State> findSolution() {
         SearchPath<State> searchPath = new SearchPath<State>();
         SearchNode<State> searchNode = this.successors.convertStateToSearchNode("start", startState, null);
@@ -25,7 +25,7 @@ public class Search<State> {
 
         while (!controller.isEmpty()) {
             searchNode = controller.remove();
-            Util.debug("the removed searchnode is" + searchNode.toString());
+            Util.debug("The removed Search Node is: " + searchNode.toString());
             searchPath.append(searchNode);
             State currentState = searchNode.getState();
             if (goal.satisfied(currentState)) {
@@ -33,12 +33,12 @@ public class Search<State> {
             }
             List<SearchNode<State>> generatedSuccessors = successors.expand(searchNode);
             for (SearchNode<State> node : generatedSuccessors) {
-                Util.debug("the generated searchnode is" + node.toString());
+                Util.debug("The generated Search Node is" + node.toString());
             }
+            Util.debug("");
             controller.insertBatch(generatedSuccessors);
         }
         return null;
-
     }
 
 }
